shellcode phase 5b: move shellcode from .text section to stack - revised
It is necessary to enable execution on the stack.
This can be done at link time by adding the appropriate keyword to the ld
flags:

    LDFLAGS += -z execstack

It is also possible to use the execstack app to toggle the 'stack-required'
bit in the binary file.

We can now remove the shellcode form the .text section and
place it on the stack.