shellcode phase 3b: machine code for calling execve - relocated
Ultimately, an attacker would like to provide input the vulnerable
application which will cause the overflow. This input is
essentially a string consisting of machine code instructions
implementing the call to the execve system call with the
"/bin/sh" parameters.

The next step is to generate a machine code character string based
on the assembly code we wrote in the previous phase.
objdump tells us what the machine code should be:

    objdump -d --section=.text naive.o > naive.objdump

 line   machine code            assembly code
 ----   ------------            -------------
  21:   48 89 e6                mov    %rsp,%rsi

Note that the character buffer - shellcode - is placed in the
.text section. This is because by default, the .text section
is executable while the stack is not. Since the buffer essentially
consists of instructions to be executed we thus place them in the
.text section for the time being.
We will see in following commits how to enable execution on the
stack.

We finally get to overwrite the stack, encoding the shellcode buffer
address instead of main's return address.

Note that we have had to add the gcc -fno-stack-protector option
to disable the default stack-smashing protection (stack canaries).