shellcode phase 6: pass shellcode to vulnerable func, exploit strcpy
We can now finally exploit strcpy() to overrun the stack with shellcode
that will spawn a shell.

The general approach to doing this is as follows:
1. Identify the address on the stack to which we will want execution
   to be transferred. We refer to this address henceforth as ADDRESS.
   This is normally the strcpy() destination address.
2. Construct the shellcode with NOP padding at the beginning of the
   buffer and with ADDRESS at the end. Schematically, the shellcode
   buffer should look something like:

   "NOP NOP NOP NOP NOP NOP NOP NOP" \
   "NOP NOP NOP NOP ...            " \
   "              ...              " \
   "... NOP NOP"                     \
                                     \
               "SHELLCODE.."         \
                                     \
                           "ADDRESS" \
   "ADDRESS ADDRESS ...            " \
   "           ...          ADDRESS"

This allows some flexibility regarding the exact offset from the start
of the shellcode buffer on stack to the precise location of the
function's return address which needs to be overwritten.
The series of NOPs at the beginning also allows the jump-to address to
be more flexible, resulting in sequential NOP execution prior to
reaching and executing the shellcode itself.

While this is a plausible technique in the case of 32 bit architectures
where the stack addresses do not contain \x00 bytes and are of the form:

    0xbffff26c

on 64 bit architectures, the stack addresses are of the form:

    0x00007fffffffd4a0

consisting of two \x00 which would be interpreted by strcpy() as
a NULL terminator byte.
Hence in our example, I've had to change the above scheme a little to:

   "NOP NOP NOP NOP NOP NOP NOP NOP" \
   "NOP NOP NOP NOP ...            " \
   "              ...              " \
   "... NOP NOP"                     \
                                     \
               "SHELLCODE.."         \
                                     \
   "FFFFFFFFFFFFFF FFFFFFFFFFFFFFFF" \
   "FFFFFFFFFFFFFF      0000ADDRESS"

From here on, we require disabling of Linux ASLR (address space
layout randomization) in order to get consistent addressing of the
shellcode buffer (which needs to be hardcoded into the shellcode).

To disable ASLR, issue (as root):

    echo 0 > /proc/sys/kernel/randomize_va_space

Note, the assembly version will necessarily cause a segmentation
fault due to attempting the write a terminating NULL at the end of
"/bin/sh". The .text section doesn't have write attributes.