shellcode phase 1: calling execve()
In a stack based buffer overflow attack, one usually wants to spawn a shell.
If the vulnerable application is installed with setuid root,
an attacker could get a root shell with such an attack.

The following series of commits shows consecutive stages of developing a
full-fledged buffer overflow attack.

In order to spawn a shell, we call execve("/bin/sh", ...).
The first stage is to do so in C, and use objdump to see how the assembler
sets up the assembly code for the execve system call.

Analyze the resulting naive.o with:

    objdump -dS --section=.text naive.o > naive.objdump

This will be the basis for the next step.