shellcode phase 2: assembly implementation of execve call within main
We use the objdump analysis of the previous stage to replace the C
implementation of the call to execve() from within main() with one
in assembly.

The jmp/call trick pushes the address "/bin/bash" on the stack and is
the basis for setting up the arguments for the system call.