shellcode phase 5a: move shellcode from .text section to stack
Now that we have a NULL terminator free working shellcode, our next step
is to remove it form the .text section and place it on the stack.

However, by default, object files and executables are marked by the
loader as not requiring an executable stack. Hence, trying to
execute the shellcode off the stack results in a segmentation fault.