shellcode phase 7: pass shellcode as input to naive executable
The real deal!
We now have all that is required to demonstrate a full fledged stack
based buffer overflow attack.
The 'naive' application receives a character buffer from the user and
strcpy()'s it to one of its own.
The input.sh script produced the shellcode buffer as was previously
developed.

To demonstrate, issue:

    ./naive `./input.sh`

Note:
The naive app receives a single string, hence the shellcode cannot
contain a \x20 character (' ') - as this will be considered a separation
of arguments. For this reason, any padding and the address used for
over-writing the return address must not contain \x20.