print_stack: add application to print stack with/without GDB
The print_stack provides a graphical display of the memory layout
following the application of the shellcode on the exploited function's
stack frame:

   ret addr  oooooooo oooooooo   <address>
   $rbp      _ _ _ _ p _ _ _ _   <address>
                     p
                     p
                    ...
                     p
                     p
                     s
                     s
                     s
                    ...
                     s
                    nop          <address>
                    nop          <address>
                    nop          <address>
                    ...
   &dst      _ _ _ _nop_ _ _ _   <address>
                     x
                     x
                    ...
   $rsb      _ _ _ _ x _ _ _ _   <address>

where:
p   - padding
s   - shellcode
nop - no operation
x   - data

The stack frame is shown both for the naive application running stand
alone and for the naive application being run in gdb. gdb incurs some
extra stack frames hence the naive stack frames are slightly lower than
they would be when not running in gdb.

The input the application must be:

./print_frame <$rbp> <&dst> <$rsb>

  or

./print_frame <no gdb $rbp> <no gdb &dst> <no gdb $rsb>
              <with gdb $rbp> <with gdb &dst> <with gdb $rsb>

The values of $rbp, &dst and $rsb should be obtained by calling the
get_stack() macro in the exploited function.
This can be done by building naive with CONFGI_SHOW_STACK=y.

The print_stack application indicates the common addresses used for NOP
instructions both when running naive natively and when running it in
GDB. One of these common addresses, should be used in the final line in
the shell code which overruns the ret_addr.

Signed-off-by: Ilan Smith <ilansm@nvidia.com>